Create UTC dates mixin
